+++ test/core/pipeline/test_pipeline.py
@@ -123,3 +123,13 @@ class TestPipeline:
                 component_visits={"erroring_component": 0},
             )
         assert "Component name: 'erroring_component'" in str(exc_info.value)
+
+    def test_run(self):
+        joiner_1 = BranchJoiner(type_=str)
+        joiner_2 = BranchJoiner(type_=str)
+        pp = Pipeline()
+        pp.add_component("joiner_1", joiner_1)
+        pp.add_component("joiner_2", joiner_2)
+        pp.connect("joiner_1", "joiner_2")
+
+        _ = pp.run({"value": "test_value"})
